Add --return_all option
Default action to compute root_id was confusing with gene ontology
where no single root for all child terms exists. This maintains
the function of --root_id but when --root_id is not specified this
option overrides the default of computing a root_id.

As currently implemented the --add-category-column option requires
a root_id. It is disabled when --return_all applies.